+++ azuredevops/internal/service/graph/data_group.go
@@ -1,4 +1,4 @@
-package azuredevops
+package graph
 
 import (
 	"fmt"
@@ -8,11 +8,11 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 	"github.com/microsoft/azure-devops-go-api/azuredevops/graph"
-	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/utils"
 	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/client"
+	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/utils"
 )
 
-func dataGroup() *schema.Resource {
+func DataGroup() *schema.Resource {
 	return &schema.Resource{
 		Read: dataSourceGroupRead,
 		Schema: map[string]*schema.Schema{
